requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
